replace supervisor hooks with Writer chat hook
